+++ Tick-It-Ticket-Service/src/main/java/com/harish/TickIt/TicketService/auth/JwtAuthenticationFilter.java
@@ -38,7 +38,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter
 																											.map(r->new org.springframework.security.core.authority.SimpleGrantedAuthority(r))
 																											.collect(java.util.stream.Collectors.toSet());
 				String username= jwtUtil.getUsernameFromToken(token);
-				Integer id= jwtUtil.getEmployeeId(token);
+				Long id= jwtUtil.getEmployeeId(token);
 				
 				UserPrincipal up= new UserPrincipal();
 				up.setEmployeeId(id);
